Allow fetching data manually without AppendTask
Currently, as soon as the user scrolls to the end of the list
EndlessAdapter triggers executeAsyncTask to call code in
cacheInBackground from an AsyncTask.

This can be restrictive for codebases that already have their own
AsyncTasks written, or existing ones that cannot be modified.

With these changes, the implementor can make a call to
setRunInBackground(false) to indicate cacheInBackground() should be
called directly, rather than from the internal AsyncTask.

Also add a new Fragment to EndlessAdapterFragmentDemo, to demonstrate
the alternate usage pattern.